Added 'Dump' function to Paul-Viewer, a very cool feature to save Matplotlib scripts as... well, scripts :-)
Two ingredients were added for this: one is the 'default_files'
variable of plotscripts, a variable which, if available, will be used
to display data in the Viewer window in the absence of other input.
The other ingredient creating a copy of the current plotscript, and
appending a 'default_files' variable to it to contain the currently
selected files.

This way, a neat workflow can be used as follows:
  . User browses through his data
  . User creates/edits plotscript to customize
    the looks of his data
  . User likes what he sees, thinks "Hm, should look
    back to that later"
  . User clicks "Dump", which creates:
     1) a copy of the current plotscript at a safe location
     2) a list of the currently selected files, so User
        can reproduce the current plot later

This involves some relative path magic, which seems to work well.